+++ tests/integration/test_telegram.py
@@ -9,28 +9,35 @@ TEST_PHOTO = "https://c1.scryfall.com/file/scryfall-cards/png/front/a/d/ad98e518
 
 
 class TestTelegramClient:
-    def test_send_message(self, telegram, chat_id):
+    def test_send_message(self, telegram, any_group_id):
+        chat_id, thread = any_group_id
         resp = telegram.send_message(
             SendMessage(
-                chat_id=chat_id, text="scryfall-telegram integration test message."
+                chat_id=chat_id,
+                message_thread_id=thread,
+                text="scryfall-telegram integration test message.",
             )
         )
         resp.raise_for_status()
 
-    def test_send_photo(self, telegram, chat_id):
+    def test_send_photo(self, telegram, any_group_id):
+        chat_id, thread = any_group_id
         resp = telegram.send_photo(
             SendPhoto(
                 chat_id=chat_id,
+                message_thread_id=thread,
                 photo=TEST_PHOTO,
                 caption="scryfall-telegram integration test photo.",
             )
         )
         resp.raise_for_status()
 
-    def test_send_media_group(self, telegram, chat_id):
+    def test_send_media_group(self, telegram, any_group_id):
+        chat_id, thread = any_group_id
         resp = telegram.send_media_group(
             SendMediaGroup(
                 chat_id=chat_id,
+                message_thread_id=thread,
                 media=[
                     InputMediaPhoto(type="photo", media=TEST_PHOTO) for _ in range(3)
                 ],