+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased
+
+### Added
+
+- Support for Channel posts.
+
+### Changed
+
+- Should now reply in the thread or topic where the bot was invoked, rather than the
+  global chat.
+
+
 ## 2.3.0 - 2023-03-05
 
 ### Changed